Refactor tafsir data hooks
